Bazelize ssx tests

+++ src/v/ssx/tests/sformat_bench.cc
@@ -39,7 +39,7 @@ void run_test(Fun fun, size_t data_size) {
     std::vector<Val> vec;
     vec.reserve(data_size);
     perf_tests::start_measuring_time();
-    for (int i = 0; i < data_size; ++i) {
+    for (size_t i = 0; i < data_size; ++i) {
         vec.emplace_back(fun(fmt::runtime("{}"), identifier));
     }
     perf_tests::do_not_optimize(
